backend/requirements.txt
fastapi
uvicorn
python-multipart
moviepy==1.0.3
openai-whisper
gTTS
deep-translator
numpy<2.0.0
decorator<5.0
